- Cambia las ubicaciones de las directive a su propia carpeta - Refactoriza para que se pueda heredar la funcionalidad común de recibos.
