+++ src/app/components/cuenta-corriente-cliente-modal/cuenta-corriente-cliente-modal.component.ts
@@ -2,7 +2,7 @@ import { Component } from '@angular/core';
 import { NgbActiveModal } from '@ng-bootstrap/ng-bootstrap';
 import { CuentasCorrientesService } from '../../services/cuentas-corrientes.service';
 import { Pagination } from '../../models/pagination';
-import { ItemSelectionModalDirective } from '../../directives/busqueda-modal/item-selection-modal.directive';
+import { ItemSelectionModalDirective } from '../../directives/item-selection-modal.directive';
 import { Observable } from 'rxjs';
 
 @Component({
