+++ src/app/directives/listado.directive.ts
@@ -5,12 +5,12 @@ import { OnDestroy, OnInit, Directive } from '@angular/core';
 import {Observable, Subscription} from 'rxjs';
 import { finalize } from 'rxjs/operators';
 import { Pagination } from '../models/pagination';
-import { MensajeModalType } from './mensaje-modal/mensaje-modal.component';
+import { MensajeModalType } from '../components/mensaje-modal/mensaje-modal.component';
 import { LoadingOverlayService } from '../services/loading-overlay.service';
 import { MensajeService } from '../services/mensaje.service';
 
 @Directive()
-export abstract class ListadoBaseDirective implements OnInit, OnDestroy {
+export abstract class ListadoDirective implements OnInit, OnDestroy {
   items = [];
   isFiltersCollapsed = true;
