+++ src/app/directives/nota-credito-venta-detalle-modal-directive.ts
@@ -1,14 +1,14 @@
-import {Cliente} from '../../models/cliente';
+import {Cliente} from '../models/cliente';
 import {FormBuilder} from '@angular/forms';
 import {NgbActiveModal} from '@ng-bootstrap/ng-bootstrap';
-import {NotasService} from '../../services/notas.service';
-import {LoadingOverlayService} from '../../services/loading-overlay.service';
-import {MensajeService} from '../../services/mensaje.service';
+import {NotasService} from '../services/notas.service';
+import {LoadingOverlayService} from '../services/loading-overlay.service';
+import {MensajeService} from '../services/mensaje.service';
 import { OnInit, Directive } from '@angular/core';
-import {ClientesService} from '../../services/clientes.service';
+import {ClientesService} from '../services/clientes.service';
 import {finalize} from 'rxjs/operators';
-import {MensajeModalType} from '../mensaje-modal/mensaje-modal.component';
-import NotaCreditoDetalleModalDirective from '../nota-credito-detalle-modal/nota-credito-detalle-modal-directive';
+import {MensajeModalType} from '../components/mensaje-modal/mensaje-modal.component';
+import NotaCreditoDetalleModalDirective from './nota-credito-detalle-modal-directive';
 
 @Directive()
 export default abstract class NotaCreditoVentaDetalleModalDirective extends NotaCreditoDetalleModalDirective  implements OnInit {
