+++ src/app/directives/nota-debito-compra-detalle-modal.directive.ts
@@ -1,12 +1,12 @@
 import {NgbActiveModal} from '@ng-bootstrap/ng-bootstrap';
 import {FormBuilder, Validators} from '@angular/forms';
-import {NotasService} from '../../../../services/notas.service';
-import {LoadingOverlayService} from '../../../../services/loading-overlay.service';
-import {MensajeService} from '../../../../services/mensaje.service';
-import {NotaDebitoDetalleModalDirective} from '../../../../components/nota-debito-detalle-modal/nota-debito-detalle-modal.directive';
+import {NotasService} from '../services/notas.service';
+import {LoadingOverlayService} from '../services/loading-overlay.service';
+import {MensajeService} from '../services/mensaje.service';
+import {NotaDebitoDetalleModalDirective} from './nota-debito-detalle-modal.directive';
 import {Directive, OnInit} from '@angular/core';
-import {Proveedor} from '../../../../models/proveedor';
-import {HelperService} from '../../../../services/helper.service';
+import {Proveedor} from '../models/proveedor';
+import {HelperService} from '../services/helper.service';
 
 @Directive()
 export abstract class NotaDebitoCompraDetalleModalDirective extends NotaDebitoDetalleModalDirective implements OnInit {
