+++ src/app/directives/nota-debito-detalle-modal.directive.ts
@@ -1,10 +1,10 @@
-import {NotaDebito} from '../../models/nota';
+import {NotaDebito} from '../models/nota';
 import {FormBuilder, FormGroup, Validators} from '@angular/forms';
 import {NgbActiveModal} from '@ng-bootstrap/ng-bootstrap';
-import {NotasService} from '../../services/notas.service';
-import {LoadingOverlayService} from '../../services/loading-overlay.service';
-import {MensajeService} from '../../services/mensaje.service';
-import {TipoDeComprobante} from '../../models/tipo-de-comprobante';
+import {NotasService} from '../services/notas.service';
+import {LoadingOverlayService} from '../services/loading-overlay.service';
+import {MensajeService} from '../services/mensaje.service';
+import {TipoDeComprobante} from '../models/tipo-de-comprobante';
 import {Directive, OnInit} from '@angular/core';
 
 @Directive()
