+++ src/app/directives/nota-debito-venta-detalle-modal.directive.ts
@@ -1,11 +1,11 @@
-import {Cliente} from '../../models/cliente';
+import {Cliente} from '../models/cliente';
 import {FormBuilder} from '@angular/forms';
 import {NgbActiveModal} from '@ng-bootstrap/ng-bootstrap';
-import {NotasService} from '../../services/notas.service';
-import {LoadingOverlayService} from '../../services/loading-overlay.service';
-import {MensajeService} from '../../services/mensaje.service';
+import {NotasService} from '../services/notas.service';
+import {LoadingOverlayService} from '../services/loading-overlay.service';
+import {MensajeService} from '../services/mensaje.service';
 import {Directive, OnInit} from '@angular/core';
-import {NotaDebitoDetalleModalDirective} from '../nota-debito-detalle-modal/nota-debito-detalle-modal.directive';
+import {NotaDebitoDetalleModalDirective} from './nota-debito-detalle-modal.directive';
 
 @Directive()
 export abstract class NotaDebitoVentaDetalleModalDirective extends NotaDebitoDetalleModalDirective implements OnInit {
