+++ src/app/directives/notas-debito.directive.ts
@@ -1,21 +1,21 @@
 import { OnInit, Directive } from '@angular/core';
 import {ActivatedRoute, Router} from '@angular/router';
-import {SucursalesService} from '../../services/sucursales.service';
-import {LoadingOverlayService} from '../../services/loading-overlay.service';
-import {MensajeService} from '../../services/mensaje.service';
+import {SucursalesService} from '../services/sucursales.service';
+import {LoadingOverlayService} from '../services/loading-overlay.service';
+import {MensajeService} from '../services/mensaje.service';
 import {FormBuilder} from '@angular/forms';
 import {NotasDirective} from './notas.directive';
-import {ClientesService} from '../../services/clientes.service';
-import {UsuariosService} from '../../services/usuarios.service';
-import {AuthService} from '../../services/auth.service';
-import {ConfiguracionesSucursalService} from '../../services/configuraciones-sucursal.service';
-import {NotasService} from '../../services/notas.service';
+import {ClientesService} from '../services/clientes.service';
+import {UsuariosService} from '../services/usuarios.service';
+import {AuthService} from '../services/auth.service';
+import {ConfiguracionesSucursalService} from '../services/configuraciones-sucursal.service';
+import {NotasService} from '../services/notas.service';
 import {finalize} from 'rxjs/operators';
-import {MensajeModalType} from '../mensaje-modal/mensaje-modal.component';
+import {MensajeModalType} from '../components/mensaje-modal/mensaje-modal.component';
 import {Observable} from 'rxjs';
-import {Pagination} from '../../models/pagination';
-import {BusquedaNotaCriteria} from '../../models/criterias/busqueda-nota-criteria';
-import {ProveedoresService} from '../../services/proveedores.service';
+import {Pagination} from '../models/pagination';
+import {BusquedaNotaCriteria} from '../models/criterias/busqueda-nota-criteria';
+import {ProveedoresService} from '../services/proveedores.service';
 
 @Directive()
 export abstract class NotasDebitoDirective extends NotasDirective implements OnInit {
