+++ src/app/feature-modules/cuentas-corrientes-proveedor/components/nota-credito-compra-detalle-factura-modal/nota-credito-compra-detalle-factura-modal.component.ts
@@ -1,5 +1,5 @@
 import { Component, OnInit } from '@angular/core';
-import NotaCreditoCompraDetalleModalDirective from '../nota-credito-compra-detalle-modal/nota-credito-compra-detalle-modal-directive';
+import NotaCreditoCompraDetalleModalDirective from '../../../../directives/nota-credito-compra-detalle-modal-directive';
 import {NuevaNotaCreditoDeFactura} from '../../../../models/nueva-nota-credito-de-factura';
 import {NgbActiveModal} from '@ng-bootstrap/ng-bootstrap';
 import {FormBuilder} from '@angular/forms';
